+++ ChangeLog.md
@@ -5,6 +5,7 @@
 **Released: WiP**
 
 - Added `Subscription.folder_id`.
+  ([#25](https://github.com/davep/oldas/pull/25))
 
 ## v0.4.0